+++ passes/S009/testdata/src/a/outside_package.go
@@ -1,0 +1,31 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = schema.Schema{
+		Type:         schema.TypeList,
+		ValidateFunc: func() {},
+	}
+
+	_ = schema.Schema{
+		Type:         schema.TypeSet,
+		ValidateFunc: func() {},
+	}
+
+	_ = map[string]*schema.Schema{
+		"name": {
+			Type:         schema.TypeList,
+			ValidateFunc: func() {},
+		},
+	}
+
+	_ = map[string]*schema.Schema{
+		"name": {
+			Type:         schema.TypeSet,
+			ValidateFunc: func() {},
+		},
+	}
+}